refactor: simplify after Phase 2 review
- Extract collect_companions() generic helper for collection macros,
  replacing ~90 lines across routes_macro/tasks_macro/static_routes_macro
  with 3 one-liner delegations
- Extract project_dir() helper in app.rs to deduplicate AUTUMN_MANIFEST_DIR
  env var resolution (was copy-pasted 3 times)
- Remove TOCTOU exists() check in find_binary() — Command::new() already
  handles missing binaries with a clear error
